adjust test for compatibility with sklearn v1.7 (#102)
* adjust test for compatibility with sklearn v1.7

* update tests

* also convert sample_weights to float32 when calling sklearn

+++ src/metrax/nnx/nnx_wrapper_test.py
@@ -89,9 +89,9 @@ class NnxWrapperTest(parameterized.TestCase):
     # TODO(jiwonshin): Use `keras.metrics.MeanSquaredError` once it supports
     # sample weights.
     expected = sklearn_metrics.mean_squared_error(
-        y_true.flatten(),
-        y_pred.flatten(),
-        sample_weight=sample_weights.flatten(),
+        y_true.astype('float32').flatten(),
+        y_pred.astype('float32').flatten(),
+        sample_weight=sample_weights.astype('float32').flatten(),
     )
     # Use lower tolerance for lower precision dtypes.
     rtol = 1e-2 if y_true.dtype in (jnp.float16, jnp.bfloat16) else 1e-05
