+++ src/metrax/regression_metrics_test.py
@@ -137,9 +137,9 @@ class RegressionMetricsTest(parameterized.TestCase):
     # TODO(jiwonshin): Use `keras.metrics.MeanAbsoluteError` once it supports
     # sample weights.
     expected = sklearn_metrics.mean_absolute_error(
-        y_true.flatten(),
-        y_pred.flatten(),
-        sample_weight=sample_weights.flatten(),
+        y_true.astype('float32').flatten(),
+        y_pred.astype('float32').flatten(),
+        sample_weight=sample_weights.astype('float32').flatten(),
     )
     # Use lower tolerance for lower precision dtypes.
     rtol = 1e-2 if y_true.dtype in (jnp.float16, jnp.bfloat16) else 1e-05
@@ -179,9 +179,9 @@ class RegressionMetricsTest(parameterized.TestCase):
     # TODO(jiwonshin): Use `keras.metrics.MeanSquaredError` once it supports
     # sample weights.
     expected = sklearn_metrics.mean_squared_error(
-        y_true.flatten(),
-        y_pred.flatten(),
-        sample_weight=sample_weights.flatten(),
+        y_true.astype('float32').flatten(),
+        y_pred.astype('float32').flatten(),
+        sample_weight=sample_weights.astype('float32').flatten(),
     )
     # Use lower tolerance for lower precision dtypes.
     rtol = 1e-2 if y_true.dtype in (jnp.float16, jnp.bfloat16) else 1e-05
